Conflicts:
	src/arch/gpu/gpumemorytransfer.cpp
	src/arch/gpu/gpumemorytransfer.hpp
	src/arch/gpu/gputhread.cpp
	src/arch/gpu/gputhread.hpp
	src/arch/gpu/version_commons.am
	src/core/basethread_decl.hpp
	src/core/copydata.hpp
	src/core/copydata_decl.hpp
	src/core/nanos-int.h
	src/core/workdescriptor.cpp
	src/support/simpleallocator.cpp

+++ src/core/cachedaccelerator.hpp
@@ -28,7 +28,7 @@ using namespace nanos;
 
 
 template <class Device>
-void CachedAccelerator<Device>::configureCache( int cacheSize, System::CachePolicyType cachePolicy )
+void CachedAccelerator<Device>::configureCache( size_t cacheSize, System::CachePolicyType cachePolicy )
 {
    if ( _cache == NULL )
       _cache = NEW DeviceCache<Device>( cacheSize, NULL, this );
